Revert "Merge pull request #146 from apiaddicts/develop"
This reverts commit 95f1deed56257d41953ce6fceac6107c48209e7f, reversing
changes made to 108118f4d6b1472f8fb73402e9721dcd42c14e05.

+++ test/oap2/format/OAR016/fail-example.js
@@ -15,7 +15,7 @@ module.exports = {
               "properties": {
                 "amount": {
                   "type": "number",
-                  "format": "int32"
+                  "format": "decimal"
                 },
                 "nested": {
                   "type": "object",
